Bug 401141 - Inner jars are not signed in CBI based build
Add code to search artifacts for inner jars and signs them

signInnerJars() Takes a jar and a list containing inner jars that
need to be signed and signs them. It performs the signing by
unzipping the jar file and signs the inner-jars then re-zips the
files into a new jar using the new function repackJar().

repackJar() Creates a jar file with the contents from a directory.

Change-Id: I657888bba41cd241b165aaf654f022bbb0a2f9e3