Only include the "Contents" folder for signing
This commit modifies the signing to only include files and folders
from the "..app/Contents" folder for singing. Only the "Contents" 
folder is relevant for Mac OS code signing. In some layouts the ".app"
folder is the top-level folder containing additional resources such as
plug-ins, etc. Those resources don't need to be signed. Thus, there is
no need for them to be in the zip for signing. 

Also add a bit more log output as well as proper error reporting from
the server-side signing script if available (sent by the signer as
plain/text).
 
Change-Id: Ib0d49da65ea6d3505fb3ab411da111960f386492
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>