Fix for improved Eclipse IDE setup: m2e Checkstyle ConversionException
This solves an Error that occurs in Eclipse with the Checkstyle Eclipse
plugin, where it gets confused between the REGEXP license for Checkstyle
and a default APL (not EPL) LICENSE.txt that is on the classpath of the
Checkstyle Maven plugin org.apache.commons.beanutils.ConversionException
line 83 in header specification is not a regular expression at
com.puppycrawl.tools.checkstyle.checks.header.RegexpHeaderCheck.postProcessHeaderLines(RegexpHeaderCheck.java:160)

https://github.com/m2e-code-quality/m2e-code-quality/issues/83 has the
full background story and and further technical analysis details.

This commit is a port from https://git.opendaylight.org/gerrit/37842.

Change-Id: I4c0a9b962e42eab9df28909403efc70ad71d2308
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>