Convert relative files paths to absolute paths
It's possible to specify relative paths in configuration files. Those
are interpreted relative to a working directory. By resolving them
within PropertiesReader we ensure that they are interpreted relative to
the working directory of the signing server, which could be a different
one than what an external tool would use.

More specifically, this fixes an issue on Mac where the code signing
tool could not access a relative keychain because it was interpreting
the relative path based on a different working directory.

Change-Id: I5a1fc3114b3cab8975ce0c4ee3716fac258d2d66
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>