Bug 532722 - Add wall timeout to each mojo for remote requests
The default timeout is 0 (infinite).

jarsigner 1.1.4 introduced a "connectTimeoutMillis" parameter, but it is
not a wall timeout, only a socket timeout. It's likely that users want
to define a wall timeout, not network-specific timeouts. Hence,
connectTimeoutMillis is now deprecated in version 1.1.5 and user are
invited to use "timeoutMillis"

Early snapshots of version 1.1.5 had "connectTimeoutMillis" for
macsigner, dmg-packager... but they has been removed in favor of the
simple wall timeout parameter "timeoutMillis".

Change-Id: I25d5a3989b37df2aaf93fc645ba4ac978e0275b2
Signed-off-by: Mikael Barbero <mikael.barbero@eclipse-foundation.org>