Flatpak: Add new option 'branchAlias'
Allows creation of additional commits in the flatpak repo pointing
the same application, but under different branch names.

This is useful for, e.g., always having a "latest" branch pointing to
the lastest build even if the stream changes. Users can follow the
"latest" branch to always be moved to the latest build, or follow the
branch for a specific stream to ensure they don't get upgraded
unexpectedly.

Change-Id: I69cad1c955c725edcd3da096bbc98f027d015517
Signed-off-by: Mat Booth <mat.booth@gmail.com>